smartops/apps/organizations/serializers.py
from rest_framework import serializers
from .models import Organization

class OrganizationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Organization
        fields = ('id', 'name', 'owner', 'created_at')
        read_only_fields = ('owner', 'created_at')
