smartops/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
redis
celery
flower
django-environ
drf-yasg
gunicorn
sentry-sdk
pytest-django
